+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+  - 🐞 [@mixte/mel-components] 修复 `MelSelect` 组件启用 `filterable` 时筛选不会忽略大小写的问题
+  - 🐞 [@mixte/mel-components] 修复 `MelSelect` 组件使用 `option` 时报的控制台警告
 
 ## [v3.5.0-beta.17]
   - 📅 2025-09-09